Creates a model to act as a cache for statistics
This is required because the app is taking so long to load the homepage. It also is broken at Heroku because it is getting timout hence the Heroku rejects the connection that takes more than 30secs.
As the census data changes yearly, store in DB the calculations result seems to be a great option.